+++ assets/appFramework/react/generate/component/src/gen-name-snake.tsx
@@ -9,10 +9,24 @@ import "./gen-name-snake.css";
 
 export class GenNamePascal extends React.Component {
     /**
+     * Message displayed in the view.
+     * @type {string}
+     */
+    public message: string;
+
+    /**
+     * Creates an instance of GenNamePascal.
+     */
+    constructor() {
+        super();
+        this.message = "Hello Gen Name Human";
+    }
+
+    /**
      * Render the component.
      * @returns {JSX.Element}
      */
     public render(): JSX.Element {
-        return (<span className="gen-name-snake">Hello Gen Name Human</span>);
+        return <span className="gen-name-snake">{ this.message }</span>;
     }
 }